+++ src/pages/tutors/tutor/index.scss
@@ -1,0 +1,11 @@
+.tutor-courses {
+  &__title {
+    background-color: #f8f9f8;
+    color: #221638;
+    text-align: center;
+    margin-bottom: 30px;
+    font-weight: 800;
+    padding: 15px 0;
+    font-size: 22px;
+  }
+}
