+++ src/pages/user/MyProfile/index.scss
@@ -1,0 +1,13 @@
+.user-profile {
+  text-align: center;
+  img {
+    margin-bottom: 20px;
+  }
+  h3 {
+    font-size: 22px;
+    font-weight: 700;
+  }
+  p {
+    margin: 0;
+  }
+}
