+++ src/pages/user/MyProfile/index.tsx
@@ -1,14 +1,15 @@
 import React, { useEffect, useContext, useState, ChangeEvent } from 'react';
 
-import PageBanner from '../../components/Common/PageBanner';
+import PageBanner from '../../../components/Common/PageBanner';
 import { EscolaLMSContext } from '@escolalms/sdk/lib/react/context';
 import { useHistory } from 'react-router-dom';
 import { useCallback } from 'react';
 import Editor from 'rich-markdown-editor';
 import Image from '@escolalms/sdk/lib/react/components/Image';
 import { API } from '@escolalms/sdk/lib';
-import Layout from '../../components/_App/Layout';
+import Layout from '../../../components/_App/Layout';
 import { useTranslation } from 'react-i18next';
+import "./index.scss";
 
 type UpdateCall = (key: keyof API.UserItem, value: any) => void;
