+++ src/style/css/style.scss
@@ -2298,18 +2298,6 @@ Profile Area CSS
   }
 }
 
-.profile-courses {
-  .title {
-    background-color: #f8f9f8;
-    color: #221638;
-    text-align: center;
-    margin-bottom: 30px;
-    font-weight: 800;
-    padding: 15px 0;
-    font-size: 22px;
-  }
-}
-
 blockquote,
 .blockquote {
   overflow: hidden;
@@ -3630,21 +3618,6 @@ Go Top CSS
   .btn {
     font-size: 12px;
     padding: 4px 8px;
-  }
-}
-
-/* User Profile */
-.user-profile {
-  text-align: center;
-  img {
-    margin-bottom: 20px;
-  }
-  h3 {
-    font-size: 22px;
-    font-weight: 700;
-  }
-  p {
-    margin: 0;
   }
 }
